[cli] add check/validate/verify subcommand
